cleanup and no more default node
