+++ src/example-types/cardinality.js
@@ -1,6 +1,7 @@
 let { getField } = require('../fields')
 
 module.exports = {
+  validContext: node => node.field,
   result: ({ field }, search, schema) =>
     search({
       aggs: {
