--- src/example-types/default.js
@@ -1,3 +1,0 @@
-module.exports = {
-  validContext: () => true,
-}
