+++ src/example-types/legacy/matchStats.js
@@ -1,7 +1,31 @@
-let twoLevelMatch = require('./twoLevelMatch')
+let _ = require('lodash/fp')
+let esTwoLevel = require('./esTwoLevelAggregation').result
+let { negate } = require('../../elasticDSL')
 
 module.exports = {
-  validContext: twoLevelMatch.validContext,
-  result: (node, search) =>
-    twoLevelMatch.result({ value_type: 'stats', ...node }, search),
+  validContext: node => node.key_field && node.value_field && node.key_value,
+  result(node, search) {
+    let filter = {
+      [node.key_type || 'term']: {
+        [node.key_field]: node.key_value,
+      },
+    }
+    return esTwoLevel(
+      _.merge(
+        {
+          key_type: 'filters',
+          key_data: {
+            filters: {
+              pass: filter,
+              fail: negate(filter),
+            },
+            field: null,
+          },
+          value_type: 'stats'
+        },
+        node
+      ),
+      search
+    )
+  },
 }
